move tmdb node creation to a custom plugin
