+++ plugins/gatsby-source-tmdb/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "gatsby-source-tmdb"
+}
